MediForge-AI-v3/requirements.txt
torch==2.3.1 torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
transformers==4.45.1 accelerate bitsandbytes
gradio==4.44.1 langgraph==0.2.5 langchain-community==0.2.16 faiss-cpu==1.8.0
google-generativeai==0.8.3 tensorflowjs==4.20.0 onnx==1.17.0
